pyproject.toml
[tool.poetry]
name = "Yale_CELI_List_Parser"
version = "0.1.0"
description = ""
authors = ["donBarbos <donbarbos@proton.me>"]

[tool.poetry.dependencies]
python = "^3.8"
selenium = "^4.7.2"
fake-useragent = "^1.1.1"
webdriver-manager = "^3.8.5"
python-dotenv = "^0.21.0"

[tool.poetry.dev-dependencies]
flake8 = "^6.0.0"
flake8-bugbear = "^22.12.6"
flake8-polyfill = "^1.0.2"
flake8-quotes = "^3.3.2"
flake8-variables-names = "^0.0.5"
flake8-black = "^0.3.6"
pyflakes = "^3.0.1"
pep8-naming = "^0.13.3"
black = "^22.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line_length = 100
target_version = ["py37", "py38", "py39"]
multi-line-output = 3
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.venv
  | \.env
  | logs
  | photos
)/
'''
